fix: address AGY production review feedback
